+++ tests/lizmap/etc/conf/localconfig.d/localconfig.ini.php
@@ -1,6 +1,6 @@
 [modules]
 pgrouting.access=2
-pgrouting.installparam="srid=2154"
+pgrouting.installparam="srid=2154;postgresql_user_group=gis_group"
 
 [jResponseHtml]
 plugins = debugbar